smtp-message: split into a few files
